refactor: test tee independently from worker
