+++ tests/bootstrap/test_app_worker_seed.py
@@ -1,5 +1,4 @@
 """
-
 Unit tests for ``pg_atlas.procrastinate.app``, ``worker``, and ``seed``.
 
 SPDX-FileCopyrightText: 2026 PG Atlas contributors
@@ -9,9 +8,7 @@ SPDX-License-Identifier: MPL-2.0
 from __future__ import annotations
 
 import datetime as dt
-import sys
 from collections import Counter
-from pathlib import Path
 
 import pytest
 import pytest_mock
@@ -57,53 +54,6 @@ def test_worker_main_runs(monkeypatch: pytest.MonkeyPatch, mocker: pytest_mock.M
     run_mock.assert_called_once()
     run_arg = run_mock.call_args.args[0]
     run_arg.close()
-
-
-def test_run_with_optional_tee_writes_stdout_and_stderr(tmp_path: Path, capsys: pytest.CaptureFixture[str]) -> None:
-    from pg_atlas.procrastinate.worker import _run_with_optional_tee
-
-    tee_file = tmp_path / "worker.log"
-
-    def _emit() -> None:
-        print("stdout-line")
-        print("stderr-line", file=sys.stderr)
-
-    _run_with_optional_tee(tee_file, _emit)
-
-    captured = capsys.readouterr()
-    log_content = tee_file.read_text(encoding="utf-8")
-    assert "stdout-line" in captured.out
-    assert "stderr-line" in captured.err
-    assert "stdout-line" in log_content
-    assert "stderr-line" in log_content
-
-
-def test_run_with_optional_tee_routes_existing_logging_handlers_to_file(tmp_path: Path) -> None:
-    import logging
-
-    from pg_atlas.procrastinate.worker import _run_with_optional_tee
-
-    logger = logging.getLogger("test_run_with_optional_tee_routes_existing_logging_handlers_to_file")
-    old_propagate = logger.propagate
-    logger.propagate = False
-    handler = logging.StreamHandler(sys.stderr)
-    handler.setFormatter(logging.Formatter("%(message)s"))
-    logger.addHandler(handler)
-
-    tee_file = tmp_path / "worker.log"
-    try:
-
-        def _emit() -> None:
-            logger.info("logging-line")
-
-        _run_with_optional_tee(tee_file, _emit)
-    finally:
-        logger.propagate = old_propagate
-        logger.removeHandler(handler)
-        handler.close()
-
-    assert handler.stream is sys.stderr
-    assert "logging-line" in tee_file.read_text(encoding="utf-8")
 
 
 async def test_process_queue_marks_stalled_jobs_failed(mocker: pytest_mock.MockerFixture) -> None:
